fix: module imports and refactor some types
