+++ config/config.default.json
@@ -1,5 +1,7 @@
 {
   "version": 1,
+  "POESESSID": null,
+  "accountName": null,
   "assets": {},
   "trading": {
     "Aetheric Fossil": {
